Добавлен пробный вариант автодокументации методов и свойств. В корень проекта добавлено "Оглавление по проекту" с описанием назначения папок. Папка Stores перенесена в ./Infrastructure/Stores. Изображения, необходимые для работы приложения перенесены в папку ./Data/Resourses/Images.
