+++ Noob Coder/Infrastructure/Commands/NavigateFeedBackFormCommand.cs
@@ -1,9 +1,12 @@
 ﻿using Noob_Coder.Infrastructure.Commands.Base;
-using Noob_Coder.Stores;
+using Noob_Coder.Infrastructure.Stores;
 using Noob_Coder.ViewModels;
 
 namespace Noob_Coder.Infrastructure.Commands
 {
+    /// <summary>
+    /// Команда перехода на страницу формы обратной связи.
+    /// </summary>
     internal class NavigateFeedBackFormCommand : CommandBase
     {
         private NavigationStore _navigationStore;
