+++ Noob Coder/Infrastructure/Commands/NavigateMenuCommand.cs
@@ -1,9 +1,12 @@
 ﻿using Noob_Coder.Infrastructure.Commands.Base;
-using Noob_Coder.Stores;
+using Noob_Coder.Infrastructure.Stores;
 using Noob_Coder.ViewModels;
 
 namespace Noob_Coder.Infrastructure.Commands
 {
+    /// <summary>
+    /// Команда перехода на страницу главного меню.
+    /// </summary>
     internal class NavigateMenuCommand : CommandBase
     {
         private NavigationStore _navigationStore;
