--- Noob Coder/Models/Readme.txt
@@ -1,1 +1,0 @@
-﻿Здесь лежат основные сущности приложения: программист, комната и т.п.
