--- Noob Coder/Services/Readme.txt
@@ -1,2 +1,0 @@
-﻿Здесь лежат сервисы. Сервисы что-то делают с сущностями.
-Т.е. если у нас в приложении есть доставка еды - вероятно заниматься ей будет отдельный сервис.
