--- Noob Coder/Stores/NavigationStore.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using Noob_Coder.ViewModels.Base;
-
-namespace Noob_Coder.Stores
-{
-    internal class NavigationStore
-    {
-        public event Action CurrentViewModelChanged;
-        private ViewModelBase _currentViewModel;
-
-        public ViewModelBase CurrentViewModel
-        {
-            get => _currentViewModel;
-            set
-            {
-                _currentViewModel = value;
-                OnCurrentViewModelChanged();
-            }
-        }
-
-        private void OnCurrentViewModelChanged()
-        {
-            CurrentViewModelChanged?.Invoke();
-        }
-    }
-}
