+++ Noob Coder/ViewModels/Base/ViewModelBase.cs
@@ -4,16 +4,23 @@ using System.Runtime.CompilerServices;
 
 namespace Noob_Coder.ViewModels.Base
 {
+    /// <summary>
+    /// Базовый абстрактный класс для модели-представления.
+    /// </summary>
     internal abstract class ViewModelBase : INotifyPropertyChanged
     {
         public virtual string Title { get; set; }
         public event PropertyChangedEventHandler? PropertyChanged;
-
+        /// <summary>
+        /// Событие оповещает представления о том, что свойство модели-представления изменило значение.
+        /// </summary>
         protected virtual void OnPropertyChanged([CallerMemberName] string? propertyName = null)
         {
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
         }
-
+        /// <summary>
+        /// Метод устанавливает значение свойства модели-представления и вызывает событие OnPropertyChanged()
+        /// </summary>
         protected bool SetField<T>(ref T field, T value, [CallerMemberName] string? propertyName = null)
         {
             if (EqualityComparer<T>.Default.Equals(field, value)) return false;
