--- Noob Coder/ViewModels/Readme.txt
@@ -1,1 +1,0 @@
-﻿Это папка для класса-прослойки между представлением (окном, которое видит пользователь) и внутренней логикой приложения.
