+++ Noob Coder/Views/FeedBackFormView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="Noob_Coder.Views.FeedBackView"
+﻿<UserControl x:Class="Noob_Coder.Views.FeedBackFormView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
