+++ Noob Coder/Views/MenuView.xaml
@@ -15,33 +15,33 @@
             <RowDefinition Height="2*" />
         </Grid.RowDefinitions>
         <Image x:Name="Logo" Grid.Column="1" Height="200" Width="200" Source="{Binding LogoImagePath}"/>
-        <Button x:Name="NewGame" Style="{DynamicResource RoundedButton}"
+        <Button x:Name="NewGameMenuButton" Style="{DynamicResource RoundedButton}"
                 Grid.Row="1" Grid.Column="1" 
                 Margin="0,20,0,0" 
                 VerticalAlignment="Top"
                 Content="{Binding NewGameButtonName}"/>
-        <Button x:Name="LoadGame" Style="{DynamicResource RoundedButton}"
+        <Button x:Name="LoadGameMenuButton" Style="{DynamicResource RoundedButton}"
                 Grid.Column="1" Grid.Row="1"  
                 Margin="0,85,0,0" 
                 VerticalAlignment="Top" HorizontalAlignment="Center"
                 Content="{Binding LoadGameButtonName}"/>
-        <Button x:Name="SaveGame" Style="{DynamicResource RoundedButton}"
+        <Button x:Name="SaveGameMenuButton" Style="{DynamicResource RoundedButton}"
                 Grid.Column="1" Grid.Row="1" 
                 Margin="0,150,0,0" 
                 VerticalAlignment="Top" HorizontalAlignment="Center"
                 Content="{Binding SaveGameButtonName}"/>
-        <Button x:Name="Setting" Style="{DynamicResource RoundedButton}"
+        <Button x:Name="NavigateSettingMenuButton" Style="{DynamicResource RoundedButton}"
                 Grid.Column="1" Grid.Row="1" 
                 Margin="0,215,0,0" 
                 VerticalAlignment="Top" HorizontalAlignment="Center"
                 Content="{Binding SettingsButtonName}"/>
-        <Button x:Name="FeedBack" Style="{DynamicResource RoundedButton}"
+        <Button x:Name="NavigateFeedBackFormMenuButton" Style="{DynamicResource RoundedButton}"
                 Grid.Column="1" Grid.Row="1"
                 Margin="0,280,0,0" 
                 VerticalAlignment="Top" HorizontalAlignment="Center" 
                 Content="{Binding FeedbackButtonName}" 
                 Command="{Binding NavigateFeedBackFormCommand}"/>
-        <Button x:Name="Exit" Style="{DynamicResource RoundedButton}"
+        <Button x:Name="ExitMenuButton" Style="{DynamicResource RoundedButton}"
                 Grid.Column="1" Grid.Row="1"
                 Margin="0,345,0,0"   
                 VerticalAlignment="Top" HorizontalAlignment="Center"
