--- Noob Coder/Views/Readme.txt
@@ -1,1 +1,0 @@
-﻿Здесь лежит часть, которая отвечает за то что видит пользователь.
